Rename glob.* to lua_glob.*
